单点登录手动注销时改为退出全部已经登录的业务系统，自动注销依旧逐个退出

+++ framework/src/main/java/com/jinhe/tss/framework/web/listener/SessionDestroyedListener.java
@@ -48,6 +48,8 @@ public class SessionDestroyedListener implements HttpSessionListener {
         
         // 注销在线用户库中对应记录信息，去除登陆用户身份证card信息
         String token = OnlineUserManagerFactory.getManager().logout(appCode, sessionId);
-        Context.destroyIdentityCard(token); 
+        if(token != null) {
+        	Context.destroyIdentityCard(token);
+        }
     }
 }
